% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{appdeploy}
\alias{appdeploy}
\title{Function to deploy latest data app for users}
\usage{
appdeploy()
}
\arguments{
\item{x}{just run to deploy app}
}
\value{
deploy app
}
\description{
Function to deploy latest data app for users
}
