% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{business}
\alias{business}
\title{Function to send last credits available for each business}
\usage{
business(empresa)
}
\arguments{
\item{empresa}{tell which business to send latest data}
}
\value{
send emails
}
\description{
Function to send last credits available for each business
}
