% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{modelupdate}
\alias{modelupdate}
\title{Function to update model}
\usage{
modelupdate()
}
\arguments{
\item{x}{just run the function}
}
\value{
csv and excel file with model
}
\description{
Function to update model
}
