% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{scraper}
\alias{scraper}
\title{Function to scrape the newest data}
\usage{
scraper()
}
\arguments{
\item{x}{just run the function}
}
\value{
The newest data available
}
\description{
Function to scrape the newest data
}
